refactor: create testfiles per game mode
