+++ tests/game-mode-connect4.spec.ts
@@ -1,0 +1,5 @@
+import { test } from "./base-test";
+
+test.describe.skip("connect4 game mode", () => {
+  test("placeholder until connect4 testcases are defined", async () => {});
+});
